Remove singleton behavior of HttpConnectionManager
